TCP and UDP ports to use Integer/int instead of Short/short.

+++ midolmanj/src/main/java/com/midokura/midolman/layer4/NatMapping.java
@@ -12,23 +12,23 @@ import com.midokura.midolman.rules.NatTarget;
 public interface NatMapping {
 
     // Allocates and records a Dnat mapping.
-    NwTpPair allocateDnat(int nwSrc, short tpSrc, int oldNwDst, short oldTpDst,
+    NwTpPair allocateDnat(int nwSrc, int tpSrc, int oldNwDst, int oldTpDst,
             Set<NatTarget> nats);
 
-    NwTpPair lookupDnatFwd(int nwSrc, short tpSrc, int oldNwDst,
-            short oldTpDst);
+    NwTpPair lookupDnatFwd(int nwSrc, int tpSrc, int oldNwDst,
+            int oldTpDst);
 
     NwTpPair lookupDnatRev(
-            int nwSrc, short tpSrc, int newNwDst, short newTpDst);
+            int nwSrc, int tpSrc, int newNwDst, int newTpDst);
 
     // Allocates and records a Snat mapping.
-    NwTpPair allocateSnat(int oldNwSrc, short oldTpSrc, int nwDst, short tpDst,
+    NwTpPair allocateSnat(int oldNwSrc, int oldTpSrc, int nwDst, int tpDst,
             Set<NatTarget> nats);
 
-    NwTpPair lookupSnatFwd(int oldNwSrc, short oldTpSrc, int nwDst, short tpDst);
+    NwTpPair lookupSnatFwd(int oldNwSrc, int oldTpSrc, int nwDst, int tpDst);
 
     NwTpPair lookupSnatRev(
-            int newNwSrc, short newTpSrc, int nwDst, short tpDst);
+            int newNwSrc, int newTpSrc, int nwDst, int tpDst);
 
     // The implementation of this method should reserve and clean up resources.
     void updateSnatTargets(Set<NatTarget> targets);
